+++ packages/flutter/test/widgets/app_overrides_test.dart
@@ -5,7 +5,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'utils.dart';
+import 'route_tester.dart';
 
 Future<void> pumpApp(WidgetTester tester) async {
   await tester.pumpWidget(
