+++ packages/flutter/test/widgets/draggable_test.dart
@@ -15,8 +15,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+import 'route_tester.dart';
 import 'semantics_tester.dart';
-import 'utils.dart';
 import 'widgets_app_tester.dart';
 
 void main() {
