+++ packages/flutter/test/widgets/navigator_test.dart
@@ -15,8 +15,8 @@ import 'package:leak_tracker_flutter_testing/leak_tracker_flutter_testing.dart';
 import 'list_tile_tester.dart';
 import 'navigator_utils.dart';
 import 'observer_tester.dart';
+import 'route_tester.dart';
 import 'semantics_tester.dart';
-import 'utils.dart';
 
 @pragma('vm:entry-point')
 Route<void> _routeBuilder(BuildContext context, Object? arguments) {
