+++ packages/flutter/test/widgets/page_forward_transitions_test.dart
@@ -5,6 +5,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+import 'route_tester.dart';
 import 'widgets_app_tester.dart';
 
 class TestTransition extends AnimatedWidget {
